requirements.txt
functions-framework==3.2.0
pymsteams
pandas
numpy
pandas_gbq
google-cloud-bigquery
